Several small cleanups
A few things:
 - Rename GCD --> GCDMin
 - Use seconds instead of ticks as the unit of time
 - Add helper functions isOnCD and setCD for readability
 - Move most of the logic from sim.Spellcasting() into several TryActivateXXX() functions, improving readability and grouping similar bits of logic together
 - Fix bug where Agent.OnSpellAccepted was not being invoked

+++ tbc/items.go
@@ -127,7 +127,7 @@ type Item struct {
 	// Since we activate all items on CD, this works fine for stuff like Quags Eye.
 	// TODO: is this the best design for this?
 	Activate   ItemActivation `json:"-"` // Activatable Ability, produces an aura
-	ActivateCD int            `json:"-"` // cooldown on activation, -1 means perm effect.
+	ActivateCD float64        `json:"-"` // cooldown on activation, -1 means perm effect.
 	CoolID     int32          `json:"-"` // ID used for cooldown
 }
